Fix warning: implicit conversion changes signedness: 'int' to 'const unsigned int' [-Wsign-conversion]
visp/modules/core/src/image/vpCannyEdgeDetection.cpp:258:94: warning: implicit conversion changes signedness: 'int' to 'const unsigned int' [-Wsign-conversion]
  258 |     upperThreshold = vpImageFilter::computeCannyThreshold(I, lowerThreshold, &m_dIx, &m_dIy, m_gaussianKernelSize,
      |                      ~~~~~~~~~~~~~                                                           ^~~~~~~~~~~~~~~~~~~~
visp/modules/core/src/image/vpCannyEdgeDetection.cpp:585:84: warning: implicit conversion changes signedness: 'int' to 'size_type' (aka 'unsigned long') [-Wsign-conversion]
  585 |       const std::pair<unsigned int, float> &candidate = m_edgeCandidateAndGradient[id];
      |                                                         ~~~~~~~~~~~~~~~~~~~~~~~~~~ ^~
visp/modules/core/src/image/vpCannyEdgeDetection.cpp:691:63: warning: implicit conversion changes signedness: 'int' to 'const unsigned int' [-Wsign-conversion]
  691 |             hasFoundStrongEdge = recursiveSearchForStrongEdge(iterTest);
      |                                  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~ ^~~~~~~~
visp/modules/core/src/image/vpCannyEdgeDetection.cpp:707:42: warning: implicit conversion changes signedness: 'const int' to 'unsigned int' [-Wsign-conversion]
  707 |         unsigned int row = coordinates / nbCols;
      |                                        ~ ^~~~~~
visp/modules/core/src/image/vpCannyEdgeDetection.cpp:708:42: warning: implicit conversion changes signedness: 'const int' to 'unsigned int' [-Wsign-conversion]
  708 |         unsigned int col = coordinates % nbCols;
      |                                        ~ ^~~~~~

+++ modules/core/src/image/vpCannyEdgeDetection.cpp
@@ -1,6 +1,6 @@
 /*
  * ViSP, open source Visual Servoing Platform software.
- * Copyright (C) 2005 - 2025 by Inria. All rights reserved.
+ * Copyright (C) 2005 - 2026 by Inria. All rights reserved.
  *
  * This software is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -255,8 +255,11 @@ vpCannyEdgeDetection::detect(const vpImage<unsigned char> &I)
   float upperThreshold = m_upperThreshold;
   float lowerThreshold = m_lowerThreshold;
   if (upperThreshold < 0) {
-    upperThreshold = vpImageFilter::computeCannyThreshold(I, lowerThreshold, &m_dIx, &m_dIy, m_gaussianKernelSize,
-                                                          m_gaussianStdev, m_gradientFilterKernelSize, m_lowerThresholdRatio,
+    upperThreshold = vpImageFilter::computeCannyThreshold(I, lowerThreshold, &m_dIx, &m_dIy,
+                                                          static_cast<unsigned int>(m_gaussianKernelSize),
+                                                          m_gaussianStdev,
+                                                          static_cast<unsigned int>(m_gradientFilterKernelSize),
+                                                          m_lowerThresholdRatio,
                                                           m_upperThresholdRatio, m_filteringAndGradientType, mp_mask);
   }
   else if (m_lowerThreshold < 0) {
@@ -582,7 +585,7 @@ vpCannyEdgeDetection::performHysteresisThresholding(const float &lowerThreshold,
     std::vector<unsigned int> localMemoryEdgeCandidates;
 #endif
     for (int id = istart; id < istop; ++id) {
-      const std::pair<unsigned int, float> &candidate = m_edgeCandidateAndGradient[id];
+      const std::pair<unsigned int, float> &candidate = m_edgeCandidateAndGradient[static_cast<std::size_t>(id)];
       if (candidate.second >= upperThreshold) {
 #ifdef VISP_HAVE_OPENMP
         localMemoryEdgeCandidates.push_back(candidate.first);
@@ -688,7 +691,7 @@ vpCannyEdgeDetection::recursiveSearchForStrongEdge(const unsigned int &coordinat
           }
           else if (type_candidate == WEAK_EDGE) {
             // Checking if the WEAK_EDGE neighbor has a STRONG_EDGE neighbor
-            hasFoundStrongEdge = recursiveSearchForStrongEdge(iterTest);
+            hasFoundStrongEdge = recursiveSearchForStrongEdge(static_cast<unsigned int>(iterTest));
           }
         }
         catch (...) {
@@ -704,8 +707,8 @@ vpCannyEdgeDetection::recursiveSearchForStrongEdge(const unsigned int &coordinat
     if (m_storeListEdgePoints) {
       if (m_edgeMap.bitmap[coordinates] != var_uc_255) {
         // Edge point not added yet to the edge list
-        unsigned int row = coordinates / nbCols;
-        unsigned int col = coordinates % nbCols;
+        unsigned int row = static_cast<unsigned int>(coordinates) / static_cast<unsigned int>(nbCols);
+        unsigned int col = static_cast<unsigned int>(coordinates) % static_cast<unsigned int>(nbCols);
         m_edgePointsList.push_back(vpImagePoint(row, col));
       }
     }
